Update global component declarations and styles (#516)
* chore: remove postcss-nesting from package.json and postcss.config.js

* chore: update vue global components declarations

This commit updates the declarations of global components in the 'vue' module. It removes the empty export type and replaces it with an empty export statement. This change is necessary to align with the changes made in the 'unplugin-vue-components' plugin.

* refactor: update global component styles to use scoped CSS

This commit updates the global component styles in the 'App.vue', 'ProjectProgress.vue', 'ProjectAbout.vue', 'ToggleMark.vue', 'AppHeader.vue', 'ProjectRewardBoxForm.vue', 'ProjectModal.vue', 'Project.vue', 'BookmarkButton.vue', 'ProjectReward.vue', 'PrimaryButton.vue', and 'ProjectStatus.vue' files to use scoped CSS instead of postcss. This change improves the encapsulation and maintainability of the component styles.

* Apply automatic changes

---------

Co-authored-by: Tsuyoshi84 <Tsuyoshi84@users.noreply.github.com>